Separate out pvr_pal API.
Split from #539. Split out the palette API.
Also adjust doxygen for pvr_set_pal_format
to properly mark the warning.
